+++ lib/model/alarm_settings.dart
@@ -1,6 +1,9 @@
 import 'package:alarm/model/notification_settings.dart';
+import 'package:alarm/src/generated/platform_bindings.g.dart';
 import 'package:flutter/widgets.dart';
 
+/// [AlarmSettings] is a model that contains all the settings to customize
+/// and set an alarm.
 @immutable
 class AlarmSettings {
   const AlarmSettings({
@@ -26,6 +29,23 @@ class AlarmSettings {
     this.isEnabled = true,
   });
 
+  /// Converts from wire datatype.
+  AlarmSettings.fromWire(AlarmSettingsWire wire)
+      : id = wire.id,
+        dateTime =
+            DateTime.fromMillisecondsSinceEpoch(wire.millisecondsSinceEpoch),
+        assetAudioPath = wire.assetAudioPath,
+        notificationSettings =
+            NotificationSettings.fromWire(wire.notificationSettings),
+        loopAudio = wire.loopAudio,
+        vibrate = wire.vibrate,
+        volume = wire.volume,
+        volumeEnforced = wire.volumeEnforced,
+        fadeDuration = wire.fadeDuration,
+        warningNotificationOnKill = wire.warningNotificationOnKill,
+        androidFullScreenIntent = wire.androidFullScreenIntent;
+
+  /// Constructs an `AlarmSettings` instance from the given JSON data.
   factory AlarmSettings.fromJson(Map<String, dynamic> json) {
     NotificationSettings notificationSettings;
 
@@ -94,6 +114,23 @@ class AlarmSettings {
   ///this value should not be changed during the application
   final int snoozeLimit;
 
+  /// Converts to wire datatype which is used for host platform communication.
+  AlarmSettingsWire toWire() => AlarmSettingsWire(
+        id: id,
+        millisecondsSinceEpoch: dateTime.millisecondsSinceEpoch,
+        assetAudioPath: assetAudioPath,
+        notificationSettings: notificationSettings.toWire(),
+        loopAudio: loopAudio,
+        vibrate: vibrate,
+        volume: volume,
+        volumeEnforced: volumeEnforced,
+        fadeDuration: fadeDuration,
+        warningNotificationOnKill: warningNotificationOnKill,
+        androidFullScreenIntent: androidFullScreenIntent,
+      );
+
+  /// Returns a hash code for this `AlarmSettings` instance using
+  /// Jenkins hash function.
   ///this could be changed during processing
   final int changeableSnoozeLimit;
   final int snoozeInterval;